+++ src/map/mapobjeto.h
@@ -16,7 +16,7 @@ class mapObjeto: public Tobjeto
       * observador (o que vai definir o LOD).
       * x,z       -> posicao do objeto
       * distancia -> distancia do objeto ao observador*/
-     void Desenhar(int x, int z,GLfloat distancia);
+     void Desenhar(float x, float z,GLfloat distancia);
 
      void* modelo3d;     // Modelo com nível máximo de detalhe
      void* modeloMedio;  // Modelo com nível médio de detalhe
